Use a common component for the main page and the registration page
